Increase default image width from 80 to 400
I wasn't expecting it to make much of a difference in quality, but it
really does. I don't think we are losing out in performance by using
a bigger artwork. 400x is a common size used by the website. No other
application uses 80x. Whatever we might have lost in client-side
processing time by requesting a bigger image, we've more than gained
back by the fact that 400x images are far more likely to be cached in
the CDN.