Add --quality option to control IIIF image size
Defaults to 843x as that's the standard for artwork detail pages on our
website. This means that the image is already likely to be in the CDN.
In fact, all of these sizes are likely to be in the CDN. This is why
I'm sticking to rough quality controls, rather than allowing for
arbitrary widths.